requirements.txt
tqdm
requests
six
nltk
spacy
sacremoses
Sphinx
sphinx_rtd_theme
pillow >= 4.1.1
torch==1.3.0
torchvision==0.4.1
torchtext
git+git://github.com/jekbradbury/revtok.git
transformers
gensim
boto
annoy
urllib3==1.25.6
scipy==1.3.1
tensorflow_hub